Restore preventing accerciser from exploring itself
7efbb3ec93a5 ('Fix for bug #670415 - Accerciser isn't explorable
through the treeview') disabled the support for preventing accerciser
from exploring itself. But we really do want this support, otherwise we
get all kinds of hangs since accerciser is waiting for itself to provide
at-spi values.

To observer the accerciser interface, one needs to run another
accerciser, and each of them will be able to explore the other.