Use Gtk.TreeModelFilter to filter the view
This gets rid of all the manual filtering code in the model,
which already has problems when filtering is enabled and new
applications show up.

This means we get a self.filter in addition to the self.model.
The GtkTreeView now works with values from the filter. Most
times we still work on the whole model, so the values must be
converted using Gtk.TreeModelFilter functions.

As a major improvement this fixes the slow addChild handling
when filtering was enabled, due to a 5s sleep per child. This
made LibreOffice StarMath crawl when changing the element
category, as this replaces all the current buttons.