Don't refresh or highlight the accerciser app
Currently the accerciser entry is unselectable and insensitive,
which makes is hard to read. And while it's not selectable, you
can still focus the entry with the keybord, but that is invisble.

Therefore this get's rid of the special "try not to select
ourself" code and simply prevent refresh and highlight of the
row, if it's ourself.