Catch exceptions related to accessing the accessibleId property
It is possible for an application to be built locally using different
versions of dependencies than those of the system. As a result, we can
encounter the condition where accessibleId is amongst the properties of
the accessible object, but attempting to retrieve the value of that
property results in an error. Should that occur, Accerciser's Interface
Viewer stops populating and cannot be used to obtain even the most
basic details of the accessible object. It's safer to just catch all
errors and display '(no ID)' regardless of the source of failure.