ipython_view: Clear output on Ctrl+L
This adds the functionality to clear the old
output from the iypthon_view when pressing
Ctrl+L (with either lower- or upper-case 'L'),
just as is possible on a standalone interactive
IPython shell.

If anything was typed in the current input line,
that text and the cursor position are remembered
and restored.