treeview: Use filtered path to select correct entry
The gtk.TreeView methods called in `_selectExistingPath` operate
on `AccessibleTreeView`'s model, which is the filtered model
(`self.filter`, s. the call to `self.set_model(self.filter)` in the
`AccessibleTreeView` constructor).

Since the path to the unfiltered model (`self.model`) gets passed
as a parameter, convert this to the filtered path before using it to
select an item.

This fixes navigating to a bookmark when
"View" -> "Show applications without children"
is disabled (the default), which was otherwise using
a too large row index in the path if any hidden applications had
a lower row index than the bookmarked application.