treeview: Populate path for bookmark right away
Accessing the given path in the model can
fail if the items have not been populated yet.

Therefore, try to access the given path again
right after populating the model to the
given path.

To make sure the model has actually been
populated, process all pending events
after the call to `self.model.popToPath(path)`
since populating is done via idles
(s. `AccessibleModel#popLevel`) that may
otherwise not have been processed yet.

This makes expanding and selecting the
corresponding item in the treeview work
when clicking on a bookmark for the first
time, while it required clicking the bookmark
a second time previously if the path had not
already been populated.

(The documentation for `selectNodeAtPath`
already claimed this:
"The path does not need to exist in the model,
only in the accessibles tree. The model will get
populated accordingly.")