interface_view: Drop manual click handler that runs into exception
Clicking into the `GtkTextView` that is used to display
and edit the text via the text interface resulted in
this exception being printed on stderr:

    Traceback (most recent call last):
      File ".../accerciser/share/accerciser/plugins/interface_view.py", line 1453, in _onTextViewPressed
        self.offset_spin.set_value(iter.get_offset())
    AttributeError: '_ResultTuple' object has no attribute 'get_offset'

Retrieving the offset from the `_ResultTuple` in
`_onTextViewPressed` could be fixed as follows:

    -    self.offset_spin.set_value(iter.get_offset())
    +    self.offset_spin.set_value(iter[1].get_offset())

However, `_onTextCursorMove` is already registered
as a handler for cursor move events, and also gets
called when the cursor gets moved by clicking into
the GtkTextView using the mouse. Since
`_onTextCursorMove` already takes care of updating
the offset in the spinbox, there seems to be no
need for `_onTextViewPressed`, which was meant
to do exactly that, so just drop it.

Take over a helpful comment from its doc
to `_onTextCursorMove`.