ipython_view: Prepare for handling bg color in ANSI escape codes
Extend the `ANSI_COLOR` map from mapping an escape code
to only a foreground color to mapping it to a tuple
of foreground and background color and set the
background color in the `gtk.TextView` accordingly.

All currently handled escape codes are unrelated to
background color (and thus have the second element
in the tuple set to `None`).

This is in preparation for an upcoming commit to
handle more escape codes seen when using accerciser
on Debian testing with ipython3 8.5.0-4 (and currently
causing exceptions due to not being handled).