interface_view: Actually update value when it changes
Make updating the displayed value for the Value
interface work:

1) Register the event listener for the
"object:property-change:accessible-value"
event type as well, which is what is documented on AT-SPI
level [1] and what all implementations I looked into
(Gtk 3, Gtk 4, Qt 5, Qt 6) actually notify about.

The ATK (not AT-SPI) doc for the `Object:accessible-value` property
says:

> * Deprecated: Since 2.12. Use atk_value_get_value_and_text() to get
> * the value, and value-changed signal to be notified on their value
> * changes.

But at least from my observations, that does not seem to
be reflected on the AT-SPI layer, s. the merge request
for this commit [2] for more details.
As discussed there, support both signals to be sure.

2) Register the event listener in `populateUI`, so the
listener gets registered for the currently selected
accessible object.

With this in place, the displayed value is actually
updated e.g. when changing the value in a spinbox
in Gedit, while previously, the old value kept being
displayedin Accerciser, and only the text for the Text
interface was updated.

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/d963aff7f6977e50517fb2ffe8b31f1db142107d/atspi/atspi-event-listener.c#L463
[2] https://gitlab.gnome.org/GNOME/accerciser/-/merge_requests/25