bookmarks: Use available dialog space for bookmark list
Allow the scrolled window with the tree view of the
existing bookmarks in the "Edit bookmarks" dialog
to occupy all of the space in the dialog, rather
than giving it only the same height as the
three buttons next to it.

This avoids unnecessary unused space in the dialog and
allows showing more bookmarks at the same time
instead of just seeing a few and having to scroll
to find the desired one, in particular if there are
many bookmarks.