interface_view: Handle exception when loading icon
For the Selection interface, a warning icon along with
the text "Too many selectable children" is shown
when the accessible object has more than 50 children.

When retrieving the warning icon from the theme fails,
that triggers an exception that wasn't handled previously,
causing interfaces to not be properly displayed,
as observed on Fedora 38:

    exception: gtk-icon-theme-error-quark: Icon 'gtk-dialog-warning' not present in theme Adwaita (0)

Catch the exception and show just the text without
the icon in that case.

Originally reported as a bug against LibreOffice:
https://bugs.documentfoundation.org/show_bug.cgi?id=156609