Use gtk.Application and gtk.ApplicationWindow
gtk.Application provides means for various things,
e.g. for menu management.

Make use of it, and use gtk.ApplicationWindow for
the main window.

This is in preparation of migrating away from the
deprecated gtk.ActionGroup etc. for menu management.

Besides the GtkApplication doc [1], [2] and [3] provide
a very helpful introduction/tutorial on how to use
GtkApplication and GtkApplicationWindow, also an example
for handling menus.

Split what was previously done in the ctor of the `Main`
class across the ctor and gtk.Application callbacks
where this seems better suited.

[1] https://docs.gtk.org/gtk3/class.Application.html
[1] https://python-gtk-3-tutorial.readthedocs.io/en/latest/application.html
[2] https://developer.gnome.org/documentation/tutorials/application.html