bookmarks: Don't rely on 'notify' signal to update BookmarkStore
So far, the `BookmarkStore` was connecting to the 'notify'
signal of each bookmark to be notified when one of the bookmark
properties changes. `BookmarkStore#._onBookmarkChanged` would
then implicitly be called by the GObject signal mechanism.

`_Bookmark#_setApp` and `_Bookmark#_setPath` would even
emit a dummy 'notify' signal for that method to get called.
The actual property name was ignored in any case.

No longer connect to and rely on the 'notify' signal, but
explicitly call `BookmarkStore#._onBookmarkChanged` when
one of the bookmark's properties changes. To be able to do that,
let the `_Bookmark` keep a reference to the `BookmarkStore`
it belongs to.

This is one step in preparation of getting rid of the `_Bookmark`
class's inheritance from `gtk.Action`, which was deprecated
in Gtk 3.10. [1]

Sample scenario making use of this:

1) create a bookmark
2) select "Bookmarks" -> "Edit Bookmarks" from the menu
3) change the title of the bookmark (double-click into
   the column, then change the text)
4) close the dialog
5) verify that the new title is shown in the Bookmarks
   menu
5) close Accerciser
6) start Accerciser again
7) verify that the new title is still shown in the
  Bookmarks menu

(When just dropping the call to

    bookmark.connect('notify', self._onBookmarkChanged)

without taking care of notifying the bookmark store
in a different way, step 7 would fail.)

[1] https://docs.gtk.org/gtk3/class.Action.html