validate: Switch to GtkfileChooserNative
Switch from `GtkFileChooserDialog` [1] to
`GtkFileChooserNative` [2].

This e.g. means that when run with environment
variable `GTK_USE_PORTAL=1` set in a KDE Plasma desktop
environment with xdg-desktop-portal-kde, the Plasma
portal implementation file chooser will be used instead
of the GTK one.

This also gets rid of the use of the stock constants
deprecated in GTK 3.10 [3].

[1] https://docs.gtk.org/gtk3/class.FileChooserDialog.html
[2] https://docs.gtk.org/gtk3/class.FileChooserNative.html`
[3] https://docs.gtk.org/gtk3/const.STOCK_CANCEL.html