Add meson build support
Autotools is neither the most modern, nor the most user-friendly build
system in this day and age anymore. The build fails with a _very_
helpful "missing separator" error message if you don't have `yelp-tools`
installed for example.

This commit makes accerciser follow the rest of the GNOME world in
porting to Meson [1], a modern Python3-based build system. Configuring,
building, and installing accerciser can now be done with the following
commands (where `_build` is a randomly-picked build directory):

```
$ meson setup -Dprefix=/usr _build
$ meson compile -C _build
$ meson install -C _build
```

Note that the `macaroon` subdir was left untouched, as it isn't even
included in the autotools build.

[1]: https://mesonbuild.com/