prefs_dialog: Port from deprecated GtkAlignment
GtkAlignment has been deprecated in GTK 3.14. [1]

No longer use it as base class for the `_HighlighterView`
in the prefs dialog, but switch that one to directly
derive from GtkTable instead of being a GtkAlignment
that contains a table.

Instead of setting a padding for the GtkAlignment,
set margins for the GtkTable.