Introduce new WindowManager class to handle window/screen infos
The idea is to move handling specific to the window
system and screen to this class, to deal with details
of how that is implemented in one single class rather
than at different places.

Initially, move the handling to retrieve screen coordinates
for Gtk 4 there, originally introduced in

    commit 238db8fbb85bb2b2bdb1f36a0fd502b3236051b6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 5 18:22:13 2023 +0200

        node: Calculate screen coords for gtk4 apps

The plan is to move more Wnck-related things there
initially.

This might also be helpful when implementing ways
that don't rely on Wnck to retrieve the relevant
information in the future, e.g. in order to make this
work on Wayland.

Related: #41