treeview: Avoid unnecessary event processing
`AccessibleTreeview#selectNodeAtPath` lets
the GLib main loop process events to
ensure that the tree has been populated
before before selecting the node, added in

    commit a092dc933985fafd5b1e2cc3374c7bbc0fb2d12e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Jul 14 09:38:57 2022 +0200

        treeview: Populate path for bookmark right away

While experimenting with the "Inspect under mouse"
feature with Gtk 4 apps (where that feature doesn't work
yet), the feature would sometimes stop working altogether,
also for other apps where it works in general.
And Accerciser would not terminate properly any more
when closed, but had to be killed.

For some reason, it was still processing events here.

No longer process events until none are left, but
only until populating the tree is finished, which
is what is of interest.

To do so, add a new class member `isPopulating` to keep
track of whether the tree is currently being populated.