Let WindowManager#supportsScreenCoords take any accessible
Extend `WindowManager#supportsScreenCoords` to take
any accessible and not require that the application
object is passed and simplify the code at the (so far)
only existing call site a bit.

In practice, calling `get_toolkit_name()` and
`get_toolkit_version()` for any accessible seems fine,
at least returns the proper result for Gtk 3, Gtk 4 and
Qt apps in a quick test.

However, in the AT-SPI XML spec [1], `ToolkitName` and
`Version` are specified for `Application` only, and the
doc for the libatspi method `atspi_accessible_get_toolkit_name` [1]
also says:

> Gets the toolkit name for an AtspiAccessible object.
> Only works on application root objects.

Therefore, play it safe and retrieve the app from the
passed object first.

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/76a3661252153d5ddda2ee29c6f1522ad7fe51b1/xml/Application.xml#L9-17
[2] https://docs.gtk.org/atspi2/method.Accessible.get_toolkit_name.html