interfaces_view: Drop deprecated xalign for table button
Drop the deprecated horizontal alignment of 0.0 for the
button to jump to the table of the selected table cell
in the "Table Cell" section in the interface view.
(Glade was showing a warning that this property is deprecated.)

As a result, the text (accessible name of the table)
is now centred on the button instead of left-aligned, which
seems fine.