KWinWindowManager: Support KWin 6 as well
The KWinWindowManager introduced in the previous
commit works fine on KWin 5.

This commit takes into account changes introduced
in Plasma 6/KWin 6 and makes this work in a test
in a KDE Plasma Neon dev unstable VM as well.

1) The path for the Kwin script's DBus objects changed.

2) The `Workspace.clientList` method of the KWin API
was renamed to `Workspace.windowList` in KWin 6
in commit [1]

    commit ceac574a688569d2fc03d4718804f828a3918aed
    Author: Vlad Zahorodnii <vlad.zahorodnii@kde.org>
    Date:   Wed Aug 23 11:42:26 2023 +0300

        scripting: Rename Workspace.clientList

Therefore, rename the existing KWin script to contain
the version number 5 in it's name and add an adjusted
version to use for KWin 6 that uses the new method
name.

In order to decide what version of KWin to support,
evaluate the `KDE_SESSION_VERSION` environment variable
that KDE Plasma sets, assuming that KDE Plasma and
KWin have the same version number (or requiring to
manually set the variable in addition when starting
Accerciser with `ACCERCISER_WINDOW_MANAGER=kwin`).

[1] https://invent.kde.org/plasma/kwin/-/commit/ceac574a688569d2fc03d4718804f828a3918aed