quick_select: Port from deprecated gdk_display_get_pointer
`gdk_display_get_pointer` was deprecated in GTK/GDK 3.0. [1]
Use `gdk_device_get_position` instead as suggested.

Also, drop the obsolete workaround to manually delete
the screen. The crash described in the bug report from 2009 [2]
is not reproducible any more with an updated sample
script that runs with GTK 3 and Python 3, and a comment
from 2015 suggests that it was already fine back then.

[1] https://docs.gtk.org/gdk3/method.Display.get_pointer.html
[2] https://bugzilla.gnome.org/show_bug.cgi?id=593732