event_monitor: Drop call to deprecated gdk_window_get_pointer
`gdk_window_get_pointer` was deprecated in GTK/GDK 3.0. [1]
This call here in `EventMonitor#_onLinkMotion` is unnecessary,
as the result is unused, so just drop it altogether.

Hovering + clicking hyperlinks to jump to an accessible
is not affected by this, still works in a quick test
when the fix to make it work at all from pending merge request [2]
is in place.

[1] https://docs.gtk.org/gdk3/method.Window.get_pointer.html
[2] https://gitlab.gnome.org/GNOME/accerciser/-/merge_requests/52