KWinWindowManager: Don't fail on win title containing quotes
When a window title was containing quotes, this would
previously break the processing in KWinWindowManager,
as these were not properly escaped in the JavaScript KWin script,
then resulting in errors when parsing on the Python side.

For example, Firefox with a browser tab with an HTML file like

    <!DOCTYPE html>
    <html>
      <head>
        <title>Title with &quot;Quotes&quot; that break things - and &apos;single quotes&apos; also </title>
      </head>
      <body>
      </body>
    </html>

would have such a window title, resulting in an array entry like this:

    {
    "caption": "Title with "Quotes" that break things - and 'single quotes' also — Mozilla Firefox",
    "bufferGeometry.x": 0,
    "bufferGeometry.y": 0,
    "bufferGeometry.width": 1920,
    "bufferGeometry.height": 1044},

and parsing that would then fail with an exception like

    WARNING:root:exception in getWindowInfo: Expecting ',' delimiter: line 51 column 25 (char 1252)

Drop the custom building of the JSON string altogether and rather build
a JSON object and let `JSON.stringify` take care of converting that to
a string, including all necessary escaping.

This gives an entry like this in the resulting JSON string:

    {'caption': 'Title with "Quotes" that break things - and \'single quotes\' also — Mozilla Firefox', 'bufferGeometry.x': 0, 'bufferGeometry.y': 0, 'bufferGeometry.width': 1920, 'bufferGeometry.height': 1044},

and makes parsing and highlighting the proper object work as expected.

Related issue: #41