KWinWindowManager: Get window stacking order from KWin
Override `WindowManager#getWindowOrder` in
`KWinWindowManager` and use the stacking order information
from KWin.

Extend the KWin script that retrieves information
for all windows to include the stacking order
and build the list from that information.
(Higher `stackingOrder` means that the window
is further up.)

This helps getting rid of another case that was relying on
Wnck, with the latter not working on Wayland.

In a quick test, "Inspect under mouse" was generally working
as expected with
`export ACCERCISER_WINDOW_MANAGER=kwin`
in a KDE Plasma X11 session on Debian testing,
using gtk4-demo and the qtbase standarddialogs example
applications for testing.
(Position for the qtbase standarddialogs example is slightly
off, but that's unrelated to this commit.)

Due to another issue that will be fixed separately in an
upcoming commit, the same use case does not yet work in a KDE
Plasma Wayland session.

Related issue: #41