quick_select: Prevent accessing unassigned variable
Assign a default value of `None` to `inner_container`,
as it can otherwise happen that this the variable is
accessed further below without having been assigned.

This was breaking the "Inspect under mouse" feature
with the `KWinWindowManager` on Wayland, where an
exception would be triggered in a call to this
method from `QuickSelect#_inspectUnderMouse`
for one of the frames of the "plasmashell" app:

    Traceback (most recent call last):
      File "/usr/lib/python3/dist-packages/pyatspi/registry.py", line 209, in eventWrapper
        return callback(event)
               ^^^^^^^^^^^^^^^
      File "/home/michi/temp/accerciser/lib/python3/dist-packages/accerciser/hotkey_manager.py", line 73, in _accEventKeyPressed
        handled = self.hotkeyPress(event.hw_code, event.modifiers)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/home/michi/temp/accerciser/lib/python3/dist-packages/accerciser/hotkey_manager.py", line 99, in hotkeyPress
        callback()
      File "/home/michi/temp/accerciser/share/accerciser/plugins/quick_select.py", line 103, in _inspectUnderMouse
        acc = self._getComponentAtCoords(frame, x, y)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/home/michi/temp/accerciser/share/accerciser/plugins/quick_select.py", line 199, in _getComponentAtCoords
        if inner_container == parent:
           ^^^^^^^^^^^^^^^
    UnboundLocalError: cannot access local variable 'inner_container' where it is not associated with a value

With this fix in place, the "Inspect under mouse"
feature works now to select the object under mouse
for example with the qtbase standarddialogs example
in a Plasma Wayland session on Debian testing when
enabling the use of the `KWinWindowManager` via
environment variable
`ACCERCISER_WINDOW_MANAGER=kwin`.

Related issue: #41