win manager: Refactor window info retrieval
Refactor the handling to get a `WindowInfo` (i.e.
corresponding information on the system/Wnck window)
for a given accessible toplevel for the Wnck-based
implementation in the `WindowManager` class:

Introduce a new method `WindowManager#getWindowInfos`
that returns a list of `WindowInfo`, containing infos
for all system windows. Use that to implement
`WindowManager#getWindowInfo` rather than dealing
with the Wnck.window class in that method.

This limits the use of Wnck-specific classes
to the new `WindowManager#getWindowInfos` method
and prepares for reuse of `WindowManager#getWindowInfo`
for deriving classes to avoid code duplication.

Related issue: #41