WindowInfo: Add stacking order
Add `stacking_index` as a new member to the
`WindowInfo` class. This is used to distinguish
which of the windows is further up on the windows
stack. A lower stacking index means that the window
is lower in the stack.

The KWin implementation already retrieves the info
via the KWin script.

For the Wnck implementation, use `Wnck.Screen.get_windows_stacked`
in `WindowManager#getWindowInfos` as well and just increase the
index for each window.

This will allow to unify `WindowManager#getWindowOrder`
and the implementation in the `KWinWindowManager` subclass
in an upcoming commit.

Related issue: #41