Unify WindowManager implementations further, consider size
Rework window info handling in `WindowManager` and its
`KWinWindowManager` subclass a bit:

1) Merge the logic from `KWinWindowManager#_getKWinWindowData` and
`KWinWindowManager#getWindowInfo` (that returns info about a single
window for a given accessible toplevel)
to override the base class's `getWindowInfos` (introduced
in a previous commit, returns infos for all system windows)
instead and drop the `KWinWindowManager#getWindowInfo` override,
so that the base class implementation is used, which was
prepared for that in the previous commit.

This means that the logic to distinguish between
two windows that have the same name by also taking
into account the corresponding sizes, originally introduced
for the Wnck-based implementation in

    commit 0860b06a1a974d96c9c0f07e71a3bea4c8c16a59
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 21 16:51:04 2023 +0100

        node: More reliably detect the right gtk4 window

now also applies for the KWin implementation.

In a test on Debian testing with Plasma 5 Wayland with
`export ACCERCISER_WINDOW_MANAGER=kwin` and gtk4-demo
or two instances of the qtbase standarddialogs sample app
running, with both windows having different sizes,
the correct one is highlighted for each one of them now
when selecting one of its objects in the a11y treeview in
Accerciser, while that was not the case yet without
this change in place.
(In a Plasma X11 session, this still doesn't work;
likely there's a mismatch between the sizes reported by
KWin and via AT-SPI.)

2) Merge the logic to sort windows by stacking index
from `KWinWindowManager#getWindowOrder` into
`KWinWindowManager#getWindowOrder`, replacing the
previous Wnck-specific implementation in the former.

This makes use of the `stacking_index` member
added to the `WindowInfo` class in the previous
commit.

Related issue: #41