WindowInfo: Add info whether on current workspace
Add an `on_current_workspace` member to the
`WindowInfo` class that describes whether the
window is on the current workspace/virtual desktop
or not.

Retrieve the corresponding info from Wnck
for `WindowManager` and via the KWin API
for `KWinWindowManager`.

The `Wnck.Screen.get_active_workspace` [1] and
`Window.html#Wnck.Window.get_workspace` [2] docs
both metion cases where the returned workspace
can be `None`, so treat the window as on the current
workspace in that case to be safe, e.g. cover this
case mentioned in Window.get_workspace doc [2]:

> Gets the current workspace self is on. If the window is pinned (on all
> workspaces), or not on any workspaces, None may be returned.

In the Plasma 5 script, just integers are available,
and only a single desktop number is returned for the
desktop the window is on.
(Since it's possible to set a window on an arbitrary
set of virtual desktops, this does not guarantee the
correct result when manually setting a window to be
on desktops 1, 2 and 4, but not all desktops, so
that particular scenario should be avoided when
using Accerciser for analysis.)

For KWin 6 (tested in a Fedora RawHide VM), the
`desktops` property is a list of `KWin::VirtualDesktop` [3],
so all virtual desktops can be retrieved that way, making
the specific scenario of a window shown on multiple but not
all desktops work as well.

The new info will be used in an upcoming commit.

[1] https://lazka.github.io/pgi-docs/Wnck-3.0/classes/Screen.html#Wnck.Screen.get_active_workspace
[2] https://lazka.github.io/pgi-docs/Wnck-3.0/classes/Window.html#Wnck.Window.get_workspace
[3] https://develop.kde.org/docs/plasma/kwin/api/#kwinwindow