Only consider windows in current workspace for "Inspect under mouse"
When using the "Inspect under mouse" feature, only consider those
windows as candidates that are on the current workspace/virtual
desktop.

It's possible that another window on a different virtual desktop
has a higher stacking index but is not even currently visible,
and an object in that window could previously be identified as
the alleged object under the mouse pointer if the window was at
the corresponding coordinates on the other workspace/virtual desktop,
instead of determining the actual window under the mouse on the current
workspace/virtual desktop.

This was observed with both, the Wnck-based `WindowManager`
in a Plasma 5 X11 session on Debian testing, and with the
KWin-based `KWinWindowManager` (enabled via
`export ACCERCISER_WINDOW_MANAGER=kwin`) in a Plasma 5 Wayland
session on Debian testing.

To avoid that, only take into account those windows that are on the
current workspace, using the `on_current_workspace` member
of the `WindowInfo`, introduced in the previous commit.