Add GNOME Shell extension and win manager using Mutter win pos
Similar to the `KWinWindowManager` implementation added
in previous commits that retrieves window information
(and more) from KWin, this adds a new `GnomeShellWindowManager`
that retrieves window information from GNOME Shell/Mutter,
without having to rely on Wnck, with the latter not working
correctly on Wayland.

Since the information cannot be directly queried from
the Python side, this adds a GNOME Shell extension
(ID: "accerciser@accerciser.gnome.org") that implements
a DBus service that provides a "GetWindowInfos" method
that returns the window information as a JSON string,
basically using the same format that the KWin scripts
(`kwin6-retrieve-window-infos.js` for KWin 6)
output.
The `GnomeShellWindowManager` than uses that
DBus interface to query the information.

The extension retrieves the information from
Mutter, which "implements the compositor side of
the Wayland protocol, while in X11 sessions it
serves as the window manager and compositing manager." [1]
As described at [1], the `Meta` import provides
access to windows and more for GNOME Shell extensions.

Helpful documentation:

* Documentation for writing GNOME Shell JavaScript (GJS)
  extensions: [2] (and other pages linked from there)
* Guide on how to use DBus from GJS: [3]
* API doc for window-related GJS classes: [4] [5]

This commit implements retrieving window position and size,
by overriding `WindowManager#getWindowInfos` to use the
information retrieved from the GNOME Shell extension.

Using this generally makes highlighting the correct area on
screen work when selecting an object in Accerciser's treeview
of an application's a11y tree.

In a test in a GNOME Wayland session on Fedora RawHide
(GNOME 46, the only supported version for now),
the exact area of a control is highlighted e.g. for gedit,
but it is some pixels off for e.g. the Qt 6 qtbase
standarddialogs sample app.

For now, the Wnck-based `WindowManager` remains the default.
Use of the new `GnomeShellWindowManager` can explicitly be
enabled as follows:

1) copy or symlink GNOME Shell extension to user's GNOME Shell
   extension directory:

    mkdir -p ~/.local/share/gnome-shell/extensions/
    ln -s $(readlink -f ./gnome-shell-extension/accerciser@accerciser.gnome.org) ~/.local/share/gnome-shell/extensions/

2) log out of GNOME session and log in again
   (needed for the extension to become available after step 1)

3) enable extension:

    gnome-extensions enable accerciser@accerciser.gnome.org

4) set environment variable:

    export ACCERCISER_WINDOW_MANAGER=gnomeshell

5) start Accerciser

After using Accerciser, the extension can be disabled
again using

    gnome-extensions disable accerciser@accerciser.gnome.org

[1] https://gjs.guide/extensions/overview/architecture.html
[2] https://gjs.guide/extensions/development/creating.html
[3] https://gjs.guide/guides/gio/dbus.html
[4] https://gjs-docs.gnome.org/meta14~14-windowactor/
[5] https://gjs-docs.gnome.org/meta14~14/meta.window

Issue: #41